% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer_systems.R
\docType{data}
\name{cer_systems}
\alias{cer_systems}
\title{Calculated annual grid imports and exports as a function of installed solar and battery capacity for 155 households}
\format{
A data frame with 114855 rows and 6 variables:
\describe{
\item{housecode}{CER household code}
\item{solar_capacity}{installed solar capacity 0-50kW in increments of 0.5kW}
\item{battery_capacity}{installed battery capacity 0-100kW in increments of 2.5KWh}
\item{demand}{annual demand in kWh (survey year 2010)}
\item{imports}{calculated total imports in kWh}
\item{exports}{calculated total exports in kWh}
#' }
}
\source{
CER, pvWatts (NREL)
}
\usage{
cer_systems
}
\description{
solar PV + battery system import and export table for 155 CER control group households. Used to find the financially optimal system.
}
\details{
derived from CER 2010 dataset
}
\keyword{datasets}
