% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{electricity_price_fun}
\alias{electricity_price_fun}
\title{electricity_price_fun}
\usage{
electricity_price_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
price per kWh in euros
}
\description{
actual path of electricity prices. For inflation expectations see electricity_price_inflation_fun
}
