% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{get_rooftop_solar_potential}
\alias{get_rooftop_solar_potential}
\title{get_rooftop_solar_potential}
\usage{
get_rooftop_solar_potential(
  floor_area,
  housing_type,
  bedrooms,
  roof_floor_ratio = 1.3,
  usable_roof_fraction = 0.55,
  kW_per_m2 = 0.15
)
}
\arguments{
\item{floor_area}{floor area in m2 from CER survey}

\item{housing_type}{housing type from CER survey 1= apartment, 5=bungalow, 2:4 = houses}

\item{bedrooms}{number of bedrooms}

\item{roof_floor_ratio}{area roof relative to footprint of house}

\item{usable_roof_fraction}{fraction of rooftop usable for solar PV}

\item{kW_per_m2}{kWp per m2 (technical parameter)}
}
\value{
potential installed solar capacity kWp.
}
\description{
estimate the rofftop solar potential
}
