% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{get_sys_npv_0}
\alias{get_sys_npv_0}
\title{get_sys_npv_0}
\usage{
get_sys_npv_0(
  params,
  demand,
  imports,
  exports,
  solar_capacity,
  battery_capacity,
  include_grant = T,
  customer = "domestic"
)
}
\arguments{
\item{params}{cost parameter long-format dataframe}

\item{demand}{annual power demand in kWh e.g. 5000 kWh}

\item{imports}{annual grid imports in kWh}

\item{exports}{annual exports to grid in kWh}

\item{solar_capacity}{installed solar PV capacity in kW}

\item{battery_capacity}{installed battery capacity in kWh}

\item{include_grant}{include SEAI grant T/F}

\item{customer}{"domestic" only at the moment}
}
\value{
npv in euros
}
\description{
same as get_sys_npv but returns a scalar only (npv)
}
