% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{get_sys_util_0}
\alias{get_sys_util_0}
\title{get_sys_util_0}
\usage{
get_sys_util_0(
  params,
  demand,
  imports,
  exports,
  solar_capacity,
  battery_capacity,
  include_grant = T,
  customer = "domestic"
)
}
\arguments{
\item{params}{cost parameter long-format dataframe}

\item{demand}{kWh}

\item{imports}{kWh}

\item{exports}{kWh}

\item{solar_capacity}{0..50 in steps of 0.5kW}

\item{battery_capacity}{0..100 in steps of 2.5 kWh}

\item{include_grant}{True or False}

\item{customer}{"domestic" is only option at the moment}
}
\value{
scalar financial partial utility
}
\description{
the financial utility of a household solar PV system. This is defined as (NPV_0-NPV)/NPV_0 where NPV_0 is the net present value
of expected future bills (including inflation expectations, discounted out to a term) and NPV is the net present value of all future cashflows
when a solar PV investment is made (amortisation, feedin tariff revenues, reduced electricity bills, tax)
}
