% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_uptake}
\alias{get_uptake}
\title{get_uptake}
\usage{
get_uptake(abm, groupvars = NULL)
}
\arguments{
\item{abm}{output of runABM}

\item{groupvars}{grouping variables (in addition to simulation and date)}
}
\value{
fraction of households that have installed solar, fraction that have installed solar & battery and storage attachment rate (ssar)
}
\description{
the fraction of households that have adopted solar by date. Also returns that fraction of households that have adopted a solar+battery system and the storage attachement rate
}
